+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Window title according to content
 - Brokers view
 - Added click filter on tasks base column
+- Added policy information to node list and node detail view ([#11](i11)).
 - Pause / Resume auto refresh button on top bar
 - Check HTTP errors and open settings modal if so [#7](i7)
 
@@ -20,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed a bug that was preventing settings to be fully updated
-- Render node list row even when facts object is not defined [#10][i10]
+- Render node list row even when facts object is not defined ([#10](i10))
+- Filter by state.installed in node list view ([#9](i9))
 
 ## [0.4.0] - 2020-11-04
 
@@ -77,5 +79,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 <!-- External links -->
+
 [i7] https://github.com/Wiston999/razorboard/issues/7
+
+[i9] https://github.com/Wiston999/razorboard/issues/9
+
 [i10] https://github.com/Wiston999/razorboard/issues/10
+
+[i11] https://github.com/Wiston999/razorboard/issues/11